Auth delegated web example for browser keysigner
 - Added the browser keysigner example in the web/ instead of client/
   (old directory)
