--- examples/sdk/auth-delegated/client/src/pages/wallet.tsx
@@ -1,47 +1,0 @@
-import { DfnsError } from '@dfns/sdk'
-import React, { FormEvent, useState } from 'react'
-import { useNavigate } from 'react-router-dom'
-
-import '../globals.css'
-import { api } from '../api'
-import useAuth from '../hooks/useAuth'
-
-export default function WalletNew(): JSX.Element {
-  const { token, keyPair } = useAuth()
-  const [submitting, setSubmitting] = useState<boolean>(false)
-  const [error, setError] = useState<DfnsError | undefined>(undefined)
-  const navigate = useNavigate()
-
-  const handleCreate = (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault()
-
-    const formData = new FormData(event.currentTarget)
-    setSubmitting(true)
-
-    api
-      .createWallet(formData.get('network') as string, token, keyPair)
-      .then(() => navigate('/'))
-      .catch((err) => {
-        setError(err)
-        setSubmitting(false)
-      })
-  }
-
-  return (
-    <form onSubmit={handleCreate}>
-      <div className="w-full">
-        <h1 className="text-2x">Create Wallet</h1>
-
-        <div className="flex items-center gap-2">
-          <input className="input" id="network" name="network" placeholder="network" />
-
-          <button className="btn" disabled={submitting} type="submit">
-            Submit
-          </button>
-        </div>
-
-        {!!error && <div className="text-red-700">{JSON.stringify(error.message)}</div>}
-      </div>
-    </form>
-  )
-}
