+++ examples/sdk/auth-delegated/server/.env.example
@@ -1,4 +1,3 @@
-<<<<<<< Updated upstream
 DFNS_API_URL='https://api.dfns.ninja'
 DFNS_APP_ID='ap-gcnp5...<redacted>...'
 DFNS_APP_ORIGIN='http://localhost:8000'
