+++ examples/sdk/auth-delegated/web/src/App.tsx
@@ -1,4 +1,4 @@
-import { BrowserRouter, Link, Route, Routes } from 'react-router-dom'
+import { BrowserRouter, Route, Routes } from 'react-router-dom'
 
 import './globals.css'
 import { AppContextProvider } from './hooks/useAppContext'
